wip ui v2
- basic mock:
  __________________________________________
  | V81 > Todo logo > list name > todo item |
  |_________________________________________|
   _________   _________   _________
  |         | |         | |         |
  |_________| |_________| |_________|
  __________________________________________
  | BOTTOM NAV | ITEM |                     |
  |_________________________________________|
