+++ app/src/pages/mod.rs
@@ -1,9 +1,3 @@
-mod about;
 mod index;
-mod my_todo_lists;
-mod todo_list;
 
-pub use about::*;
 pub use index::*;
-pub use my_todo_lists::*;
-pub use todo_list::*;
