+++ entity/src/prelude.rs
@@ -1,3 +1,3 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.0
 
-pub use crate::todos::Entity as Todo;
+pub use crate::{lists::Entity as List, todos::Entity as Todo};
